fix: seperate host and wazero runtime
